@@ -1,84 +1,87 @@
 <script lang="ts">
-  import { createEventDispatcher } from "svelte";
-  import { supabase } from "../supabaseClient"
+  import { supabase } from "../supabaseClient";
 
-  export let size: number
-  export let url: string
-  
-  let avatarUrl: string = null
-  let uploading = false
-  let files: FileList
+  interface Props {
+    size: number;
+    url?: string | null;
+    onupload?: () => void;
+  }
 
-  const dispatch = createEventDispatcher()
+  let { size, url = $bindable(null), onupload }: Props = $props();
+
+  let avatarUrl = $state<string | null>(null);
+  let uploading = $state(false);
+  let files = $state<FileList>();
 
   const downloadImage = async (path: string) => {
     try {
-      const { data, error } = await supabase.storage.from('avatars').download(path)
+      const { data, error } = await supabase.storage
+        .from("avatars")
+        .download(path);
 
       if (error) {
-        throw error
+        throw error;
       }
 
-      const url = URL.createObjectURL(data)
-      avatarUrl = url
+      const url = URL.createObjectURL(data);
+      avatarUrl = url;
     } catch (error) {
       if (error instanceof Error) {
-        console.log('Error downloading image: ', error.message)
+        console.log("Error downloading image: ", error.message);
       }
     }
-  }
+  };
 
   const uploadAvatar = async () => {
     try {
-      uploading = true
+      uploading = true;
 
       if (!files || files.length === 0) {
-        throw new Error('You must select an image to upload.')
+        throw new Error("You must select an image to upload.");
       }
 
-      const file = files[0]
-      const fileExt = file.name.split('.').pop()
-      const filePath = `${Math.random()}.${fileExt}`
+      const file = files[0];
+      const fileExt = file.name.split(".").pop();
+      const filePath = `${Math.random()}.${fileExt}`;
 
-      let { error } = await supabase.storage
-        .from('avatars')
-        .upload(filePath, file)
-      
+      const { error } = await supabase.storage
+        .from("avatars")
+        .upload(filePath, file);
+
       if (error) {
-        throw error
+        throw error;
       }
 
-      url = filePath
-      dispatch('upload')
+      url = filePath;
+      onupload?.();
     } catch (error) {
       if (error instanceof Error) {
-        alert(error.message)
+        alert(error.message);
       }
     } finally {
-      uploading = false
+      uploading = false;
     }
-  }
+  };
 
-  $: if (url) downloadImage(url)
+  $effect(() => {
+    if (url) downloadImage(url);
+  });
 </script>
 
 <div style="width: {size}px" aria-live="polite">
-  {#if avatarUrl}  
+  {#if avatarUrl}
     <img
       src={avatarUrl}
-      alt={avatarUrl ? 'Avatar' : 'No image'}
+      alt={avatarUrl ? "Avatar" : "No image"}
       class="avatar image"
-      style="height: {size}px; width: {size}px"
+      style="height: {size}px, width: {size}px"
     />
   {:else}
-    <div
-      class="avatar no-image"
-      style="height: {size}px; width: {size}px"
-    />
+    <div class="avatar no-image" style="height: {size}px, width: {size}px"></div>
   {/if}
   <div style="width: {size}px">
     <label class="button primary block" for="single">
-      {uploading ? 'Uploading ...' : 'Upload avatar'}
+      {uploading ? "Uploading ..." : "Upload avatar"}
     </label>
     <span style="display:none">
       <input
@@ -86,7 +89,7 @@
         id="single"
         accept="image/*"
         bind:files
-        on:change={uploadAvatar}
+        onchange={uploadAvatar}
         disabled={uploading}
       />
     </span>
