@@ -39,7 +39,7 @@ const TableNode = ({
   return (
     <>
       {data.isForeign ? (
-        <header className="text-[0.55rem] px-2 py-1 border-[0.5px] rounded-[4px] bg-alternative text-default flex gap-1 items-center">
+        <header className="text-[0.55rem] px-2 py-1 border-[0.5px] rounded-[4px] bg-alternative flex gap-1 items-center">
           {data.name}
           {targetPosition && (
             <Handle
@@ -57,7 +57,7 @@ const TableNode = ({
         >
           <header
             className={cn(
-              'text-[0.55rem] pl-2 pr-1 bg-alternative text-default flex items-center justify-between',
+              'text-[0.55rem] pl-2 pr-1 bg-alternative flex items-center justify-between',
               itemHeight
             )}
           >
