@@ -171,8 +171,8 @@ const StorageSettings = () => {
                 </div>
 
                 <div className="relative flex flex-col col-span-12 gap-x-6 gap-y-2 lg:col-span-8">
-                  <div className="grid grid-cols-12 col-span-12 gap-2 items-start">
-                    <div className="col-span-8">
+                  <div className="flex items-center gap-2">
+                    <div className="flex-grow">
                       <FormField_Shadcn_
                         control={form.control}
                         name="fileSizeLimit"
@@ -194,7 +194,7 @@ const StorageSettings = () => {
                         )}
                       />
                     </div>
-                    <div className="col-span-4">
+                    <div className="flex-shrink-0">
                       <FormField_Shadcn_
                         control={form.control}
                         name="unit"
