@@ -1,11 +1,11 @@
 <script lang="ts">
-  import { onMount } from 'svelte';
-  import { supabase } from './supabaseClient';
-  import type { AuthSession } from '@supabase/supabase-js';
-  import Account from './lib/Account.svelte';
+  import { onMount } from 'svelte'
+  import { supabase } from './supabaseClient'
+  import type { AuthSession } from '@supabase/supabase-js'
+  import Account from './lib/Account.svelte'
   import Auth from './lib/Auth.svelte'
 
-  let session: AuthSession
+  let session = $state<AuthSession | null>(null)
 
   onMount(() => {
     supabase.auth.getSession().then(({ data }) => {
@@ -20,8 +20,8 @@
 
 <div class="container" style="padding: 50px 0 100px 0">
   {#if !session}
-    <Auth />
+  <Auth />
   {:else}
-    <Account {session} />
+  <Account {session} />
   {/if}
 </div>
