@@ -5,7 +5,7 @@ import './Input.css'
 export const SIZES = ['small', 'medium']
 
 /**
- * @deprecated Use ./Input_shadcn_ with type="number" instead or ./ui-patterns/data-inputs/input with type="number"
+ * @deprecated Use `import { Input_shadcn_ } from "ui"` with `type="number"` instead or ./ui-patterns/data-inputs/input with `type="number"`
  */
 const Input = ({
   className = '',
