@@ -38,7 +38,7 @@ interface TabsSubComponents {
 }
 
 /**
- * @deprecated Use ./Tabs_shadcn_ instead
+ * @deprecated Use `import { Tabs_shadcn_ } from "ui"` instead
  */
 const Tabs: React.FC<PropsWithChildren<TabsProps>> & TabsSubComponents = ({
   defaultActiveId,
